fix(inspector) mouseover effect and missing contextmenu
