+++ editor/src/components/context-menu-wrapper.tsx
@@ -199,6 +199,24 @@ export class InspectorContextMenuWrapper<T> extends ReactComponent<ContextMenuWr
         css={{
           width: '100%',
           ...this.props.style,
+          '--control-styles-interactive-unset-main-color': UtopiaTheme.color.fg7.value,
+          '--control-styles-interactive-unset-secondary-color': UtopiaTheme.color.fg7.value,
+          '--control-styles-interactive-unset-track-color': UtopiaTheme.color.bg5.value,
+          '--control-styles-interactive-unset-rail-color': UtopiaTheme.color.bg3.value,
+          '&:hover': {
+            '--control-styles-interactive-unset-main-color': getControlStyles('simple').mainColor,
+            '--control-styles-interactive-unset-secondary-color': getControlStyles('simple')
+              .secondaryColor,
+            '--control-styles-interactive-unset-track-color': getControlStyles('simple').trackColor,
+            '--control-styles-interactive-unset-rail-color': getControlStyles('simple').railColor,
+          },
+          '&:focus-within': {
+            '--control-styles-interactive-unset-main-color': getControlStyles('simple').mainColor,
+            '--control-styles-interactive-unset-secondary-color': getControlStyles('simple')
+              .secondaryColor,
+            '--control-styles-interactive-unset-track-color': getControlStyles('simple').trackColor,
+            '--control-styles-interactive-unset-rail-color': getControlStyles('simple').railColor,
+          },
         }}
       >
         <React.Fragment>
